Updating code based on PR #174 comments
